src/gta/velocity_ocr_piprequirements.txt
numpy
pillow
pytesseract
tqdm
scikit-image
scikit-learn
scipy